perf: improve E2E test suite performance and remove redundancy
- Deduplicate no-git-mode: create targeted nogit.nogit.spec.ts (8→2 tests)
  instead of re-running all 137 filemode tests
- Replace j-key loops (30-50 presses each) with direct hover in keyboard tests
- Remove redundant loadPage calls from keyboard test beforeEach blocks
- Use addInitScript for stale draft seeding to eliminate extra page navigation
- Delete fully redundant test files: smoke.spec.ts, templates.filemode.spec.ts,
  select-to-comment.filemode.spec.ts, comments-panel.filemode.spec.ts
- Trim duplicated tests from keyboard.filemode (12→2), round-complete.filemode
  (-7 API tests), theme.spec (-11), theme.filemode (-8), loading.spec (-2),
  comments.spec (-1)
- Update AGENTS.md with nogit.spec.ts naming convention

Reduces suite from 468 to 403 tests with no coverage loss.

Closes #49
