+++ e2e/tests/theme.filemode.spec.ts
@@ -71,110 +71,3 @@ test.describe('Theme — File Mode', () => {
     expect(darkLeft).toBe('66.666%');
   });
 });
-
-// ============================================================
-// File Sections Tests (file mode)
-// ============================================================
-test.describe('File Sections — File Mode', () => {
-  test.beforeEach(async ({ page }) => {
-    await loadPage(page);
-  });
-
-  test('all files start expanded (none are deleted in file mode)', async ({ page }) => {
-    const sections = page.locator('details.file-section');
-    const count = await sections.count();
-    expect(count).toBe(3); // plan.md, server.go, handler.js
-
-    for (let i = 0; i < count; i++) {
-      await expect(sections.nth(i)).toHaveAttribute('open', '');
-    }
-  });
-
-  test('clicking file header toggles open/closed', async ({ page }) => {
-    const section = page.locator('.file-section').filter({ hasText: 'plan.md' });
-    await expect(section).toHaveAttribute('open', '');
-
-    // Click the summary to collapse
-    const header = section.locator('summary.file-header');
-    await header.click();
-
-    await expect(section).not.toHaveAttribute('open', '');
-
-    // Click again to re-open
-    await header.click();
-    await expect(section).toHaveAttribute('open', '');
-  });
-});
-
-// ============================================================
-// Table of Contents Tests (multi-file mode — TOC should be hidden)
-// ============================================================
-test.describe('Table of Contents — File Mode (multi-file)', () => {
-  test.beforeEach(async ({ page }) => {
-    await loadPage(page);
-  });
-
-  test('TOC toggle button is hidden when reviewing multiple files', async ({ page }) => {
-    const tocToggle = page.locator('#tocToggle');
-    await expect(tocToggle).toBeHidden();
-  });
-
-  test('TOC panel stays hidden when reviewing multiple files', async ({ page }) => {
-    const toc = page.locator('#toc');
-    await expect(toc).toHaveClass(/toc-hidden/);
-  });
-});
-
-// ============================================================
-// Finish Review Tests (file mode)
-// ============================================================
-test.describe('Finish Review — File Mode', () => {
-  test.beforeEach(async ({ page }) => {
-    await loadPage(page);
-  });
-
-  test('clicking Finish Review shows the waiting overlay', async ({ page }) => {
-    const finishBtn = page.locator('#finishBtn');
-    await expect(finishBtn).toBeVisible();
-    await expect(finishBtn).toHaveText('Finish Review');
-
-    await finishBtn.click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-  });
-
-  test('waiting overlay shows "Review Complete" text', async ({ page }) => {
-    await page.locator('#finishBtn').click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-
-    const heading = overlay.locator('h3');
-    await expect(heading).toHaveText('Review Complete');
-  });
-
-  test('clicking "Back to editing" hides the overlay', async ({ page }) => {
-    await page.locator('#finishBtn').click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-
-    await page.locator('#backToEditing').click();
-    await expect(overlay).not.toHaveClass(/active/);
-  });
-
-  test('after going back, Finish Review button is available again', async ({ page }) => {
-    const finishBtn = page.locator('#finishBtn');
-    await finishBtn.click();
-
-    await expect(page.locator('#waitingOverlay')).toHaveClass(/active/);
-
-    await page.locator('#backToEditing').click();
-    await expect(page.locator('#waitingOverlay')).not.toHaveClass(/active/);
-
-    await expect(finishBtn).toBeEnabled();
-    await expect(finishBtn).toHaveText('Finish Review');
-    await expect(finishBtn).toHaveClass(/btn-primary/);
-  });
-});
