+++ e2e/tests/theme.spec.ts
@@ -74,158 +74,3 @@ test.describe('Theme — Git Mode', () => {
     expect(darkLeft).toBe('66.666%');
   });
 });
-
-// ============================================================
-// File Sections Tests (git mode)
-// ============================================================
-test.describe('File Sections — Git Mode', () => {
-  test.beforeEach(async ({ page }) => {
-    await loadPage(page);
-  });
-
-  test('file sections are <details> elements', async ({ page }) => {
-    const sections = page.locator('details.file-section');
-    const count = await sections.count();
-    expect(count).toBeGreaterThan(0);
-  });
-
-  test('first non-deleted file starts open', async ({ page }) => {
-    // In git mode we have: plan.md (added), server.go (modified), handler.js (added), deleted.txt (deleted)
-    // Non-deleted files should start open
-    const planSection = page.locator('#file-section-plan\\.md');
-    await expect(planSection).toHaveAttribute('open', '');
-
-    const serverSection = page.locator('#file-section-server\\.go');
-    await expect(serverSection).toHaveAttribute('open', '');
-  });
-
-  test('clicking file header toggles open/closed', async ({ page }) => {
-    const section = page.locator('#file-section-plan\\.md');
-    await expect(section).toHaveAttribute('open', '');
-
-    // Click the summary to collapse
-    const header = section.locator('summary.file-header');
-    await header.click();
-
-    // Should now be collapsed (no open attribute)
-    await expect(section).not.toHaveAttribute('open', '');
-
-    // Click again to re-open
-    await header.click();
-    await expect(section).toHaveAttribute('open', '');
-  });
-
-  test('file headers show status badges', async ({ page }) => {
-    // plan.md is "added" -> badge says "New File"
-    const planBadge = page.locator('#file-section-plan\\.md .file-header-badge');
-    await expect(planBadge).toHaveText('New File');
-
-    // server.go is "modified" -> badge says "Modified"
-    const serverBadge = page.locator('#file-section-server\\.go .file-header-badge');
-    await expect(serverBadge).toHaveText('Modified');
-
-    // deleted.txt is "deleted" -> badge says "Deleted"
-    const deletedBadge = page.locator('#file-section-deleted\\.txt .file-header-badge');
-    await expect(deletedBadge).toHaveText('Deleted');
-
-    // handler.js is "added" -> badge says "New File"
-    const handlerBadge = page.locator('#file-section-handler\\.js .file-header-badge');
-    await expect(handlerBadge).toHaveText('New File');
-  });
-
-  test('file headers show diff stats with additions and deletions', async ({ page }) => {
-    // server.go has both additions and deletions
-    const serverStats = page.locator('#file-section-server\\.go .file-header-stats');
-    await expect(serverStats).toBeVisible();
-
-    const addSpan = serverStats.locator('.add');
-    await expect(addSpan).toBeVisible();
-    await expect(addSpan).toContainText('+');
-
-    const delSpan = serverStats.locator('.del');
-    await expect(delSpan).toBeVisible();
-    await expect(delSpan).toContainText('-');
-  });
-
-  test('deleted file starts collapsed', async ({ page }) => {
-    const deletedSection = page.locator('#file-section-deleted\\.txt');
-    await expect(deletedSection).toBeVisible();
-    // Deleted file should NOT have the open attribute
-    await expect(deletedSection).not.toHaveAttribute('open', '');
-  });
-});
-
-// ============================================================
-// Table of Contents Tests (git mode — TOC is hidden)
-// ============================================================
-test.describe('Table of Contents — Git Mode', () => {
-  test('TOC toggle button is hidden in git mode', async ({ page }) => {
-    await loadPage(page);
-
-    const tocToggle = page.locator('#tocToggle');
-    await expect(tocToggle).toBeHidden();
-  });
-});
-
-// ============================================================
-// Finish Review Tests (git mode)
-// ============================================================
-test.describe('Finish Review — Git Mode', () => {
-  test.beforeEach(async ({ page }) => {
-    await loadPage(page);
-  });
-
-  test('clicking Finish Review shows the waiting overlay', async ({ page }) => {
-    const finishBtn = page.locator('#finishBtn');
-    await expect(finishBtn).toBeVisible();
-    await expect(finishBtn).toHaveText('Finish Review');
-
-    await finishBtn.click();
-
-    // Waiting overlay should become active
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-  });
-
-  test('waiting overlay shows "Review Complete" text', async ({ page }) => {
-    await page.locator('#finishBtn').click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-
-    // The overlay dialog should contain "Review Complete"
-    const heading = overlay.locator('h3');
-    await expect(heading).toHaveText('Review Complete');
-  });
-
-  test('clicking "Back to editing" hides the overlay', async ({ page }) => {
-    await page.locator('#finishBtn').click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-
-    // Click back to editing
-    const backBtn = page.locator('#backToEditing');
-    await backBtn.click();
-
-    // Overlay should no longer have the active class
-    await expect(overlay).not.toHaveClass(/active/);
-  });
-
-  test('after going back, Finish Review button is available again', async ({ page }) => {
-    const finishBtn = page.locator('#finishBtn');
-    await finishBtn.click();
-
-    const overlay = page.locator('#waitingOverlay');
-    await expect(overlay).toHaveClass(/active/);
-
-    // Go back
-    await page.locator('#backToEditing').click();
-    await expect(overlay).not.toHaveClass(/active/);
-
-    // Finish button should be enabled and say "Finish Review"
-    await expect(finishBtn).toBeEnabled();
-    await expect(finishBtn).toHaveText('Finish Review');
-    await expect(finishBtn).toHaveClass(/btn-primary/);
-  });
-});
